[9.0][IMP] pos_payment_terminal: Update payment lines with response
Payment lines are updated with response details

Amount due is updated if 'success' == True in response
