+++ pos_payment_terminal/i18n/pos_payment_terminal.pot
@@ -4,7 +4,7 @@
 #
 msgid ""
 msgstr ""
-"Project-Id-Version: Odoo Server 10.0\n"
+"Project-Id-Version: Odoo Server 9.0c\n"
 "Report-Msgid-Bugs-To: \n"
 "Last-Translator: <>\n"
 "Language-Team: \n"
@@ -34,13 +34,18 @@ msgid "Payment Mode"
 msgstr ""
 
 #. module: pos_payment_terminal
+#: model:ir.model,name:pos_payment_terminal.model_pos_order
+msgid "Point of Sale"
+msgstr ""
+
+#. module: pos_payment_terminal
 #: model:ir.model.fields,help:pos_payment_terminal.field_account_journal_payment_mode
 msgid "Select the payment mode sent to the payment terminal"
 msgstr ""
 
 #. module: pos_payment_terminal
 #. openerp-web
-#: code:addons/pos_payment_terminal/static/src/xml/pos_payment_terminal.xml:6
+#: code:addons/pos_payment_terminal/static/src/xml/pos_payment_terminal.xml:10
 #, python-format
 msgid "Start transaction"
 msgstr ""
