+++ pos_payment_terminal/static/src/xml/pos_payment_terminal.xml
@@ -10,6 +10,10 @@
                    <button class="payment-terminal-transaction-start" t-att-data-cid='line.cid'>Start transaction</button>
                 </td>
             </t>
+            <t t-if="!line.cashregister.journal.payment_mode or !widget.pos.config.iface_payment_terminal">
+                <td>
+                </td>
+            </t>
         </t>
     </t>
 </templates>
