+++ pos_payment_terminal/views/pos_payment_terminal_template.xml
@@ -3,7 +3,9 @@
 
     <template id="assets" inherit_id="point_of_sale.assets" name="pos_payment_terminal assets">
         <xpath expr="." position="inside">
-            <script type="text/javascript" src="/pos_payment_terminal/static/src/js/pos_payment_terminal.js"></script>
+            <script type="text/javascript" src="/pos_payment_terminal/static/src/js/devices.js"></script>
+            <script type="text/javascript" src="/pos_payment_terminal/static/src/js/screens.js"></script>
+            <script type="text/javascript" src="/pos_payment_terminal/static/src/js/models.js"></script>
             <link rel="stylesheet" href="/pos_payment_terminal/static/src/css/pos_payment_terminal.css" />
       </xpath>
     </template>
